added related_names

+++ services/movie_session.py
@@ -50,5 +50,6 @@ def update_movie_session(
 
 
 def delete_movie_session_by_id(session_id: int) -> None:
-    movie_session = MovieSession.objects.get(id=session_id)
-    movie_session.delete()
+    movie_session = get_movies_sessions().filter(id=session_id).first()
+    if movie_session:
+        movie_session.delete()
